requirements.txt
streamlit
pandas
plotly
langgraph
langchain-core
langchain-groq
python-dotenv
openpyxl
fpdf2